+++ modules/cloudwatch_dashboard.tf
@@ -1,4 +1,5 @@
 locals {
+  metrics_namespace = "DCE/AccountPool"
   vars = {
     api_name                                = aws_api_gateway_rest_api.gateway_api.name
     api_stage_name                          = aws_api_gateway_stage.api.stage_name
@@ -15,11 +16,13 @@ locals {
     publish_lease_events_lambda_name        = module.publish_lease_events_lambda.name
     update_principal_policy_name            = module.update_principal_policy.name
     error_scraper_query                     = "fields @timestamp, @message | sort @timestamp desc | filter @message ~= \\\"error\\\" or @message ~= \\\"fail\\\" or @message ~= \\\"fork/exec\\\" | display @timestamp, @message, @logStream| limit 10"
+    metrics_namespace_var                   = local.metrics_namespace
+    account_pool_metrics_widget_period      = var.account_pool_metrics_widget_period
   }
 }
 
 resource "aws_cloudwatch_dashboard" "main" {
   count          = var.cloudwatch_dashboard_toggle == "true" ? 1 : 0
   dashboard_name = "DCE-${var.namespace}"
-  dashboard_body = templatefile("${path.module}/fixtures/dashboards/cloudwatch_dashboard.json", local.vars)
+  dashboard_body = replace(templatefile("${path.module}/fixtures/dashboards/cloudwatch_dashboard.json", local.vars), "$$$$", "$")
 }